requirements.txt
tensorflow==2.10.1
keras==2.10.0
numpy>=1.21.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
pillow>=8.3.0
pandas>=1.3.0
protobuf<3.20.0
